# python-flask/swagger_server/models/product.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Product(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, product: str=None, name: str=None, active: bool=True, company: str=None):  # noqa: E501
        """Product - a model defined in Swagger

        :param product: The product of this Product.  # noqa: E501
        :type product: str
        :param name: The name of this Product.  # noqa: E501
        :type name: str
        :param active: The active of this Product.  # noqa: E501
        :type active: bool
        :param company: The company of this Product.  # noqa: E501
        :type company: str
        """
        self.swagger_types = {
            'product': str,
            'name': str,
            'active': bool,
            'company': str
        }

        self.attribute_map = {
            'product': 'product',
            'name': 'name',
            'active': 'active',
            'company': 'company'
        }
        self._product = product
        self._name = name
        self._active = active
        self._company = company

    @classmethod
    def from_dict(cls, dikt) -> 'Product':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Product of this Product.  # noqa: E501
        :rtype: Product
        """
        return util.deserialize_model(dikt, cls)

    @property
    def product(self) -> str:
        """Gets the product of this Product.

        Auto-generated primary key field  # noqa: E501

        :return: The product of this Product.
        :rtype: str
        """
        return self._product

    @product.setter
    def product(self, product: str):
        """Sets the product of this Product.

        Auto-generated primary key field  # noqa: E501

        :param product: The product of this Product.
        :type product: str
        """
        if product is None:
            raise ValueError("Invalid value for `product`, must not be `None`")  # noqa: E501

        self._product = product

    @property
    def name(self) -> str:
        """Gets the name of this Product.


        :return: The name of this Product.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Product.


        :param name: The name of this Product.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def active(self) -> bool:
        """Gets the active of this Product.


        :return: The active of this Product.
        :rtype: bool
        """
        return self._active

    @active.setter
    def active(self, active: bool):
        """Sets the active of this Product.


        :param active: The active of this Product.
        :type active: bool
        """

        self._active = active

    @property
    def company(self) -> str:
        """Gets the company of this Product.

        This property is a reference to a Company  # noqa: E501

        :return: The company of this Product.
        :rtype: str
        """
        return self._company

    @company.setter
    def company(self, company: str):
        """Sets the company of this Product.

        This property is a reference to a Company  # noqa: E501

        :param company: The company of this Product.
        :type company: str
        """
        if company is None:
            raise ValueError("Invalid value for `company`, must not be `None`")  # noqa: E501

        self._company = company