# python-flask/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: bampli-api
  description: The API for the Business Amplifier project
  contact:
    name: Jose Motta
    url: http://bampli.com
    email: josemotta@bampli.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
servers:
- url: https://virtserver.swaggerhub.com/motta/bampli/1.0.0-oas3
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/motta/bampli/1.0.0
...
paths:
...
  /product/:
    get:
      tags:
      - Company
      summary: Load the list of Products
      description: Loads list of Products
      operationId: search_products
      parameters:
      - name: size
        in: query
        description: Size of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: integer
        x-example: 10
      - name: page
        in: query
        description: Number of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: number
        x-example: 1
      - name: sort
        in: query
        description: 'Order in which to retrieve the results. Multiple sort criteria
          can be passed. Example: sort=name ASC,city DESC'
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: name ASC
      - name: name
        in: query
        description: Allows to filter the collections of result by name
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: George Street Brewery
      - name: company
        in: query
        description: Allows to filter the collections of result by company
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: 0e8cedd0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          headers:
            X-Page-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Number:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Size:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 25
            X-Total-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 2
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
                x-content-type: application/json
              examples:
                response:
                  value: |-
                    [{
                      "product": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Intel BT200 Motherboard",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }]
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.company_controller
    post:
      tags:
      - Company
      summary: Create a new Product
      description: Adds a Product
      operationId: add_product
      requestBody:
        $ref: '#/components/requestBodies/Product'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                response:
                  value: |-
                    {
                      "product": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Intel BT200 Motherboard",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
  /product/{productid}:
    get:
      tags:
      - Company
      summary: Load an individual Product
      description: Loads a Product
      operationId: get_product
      parameters:
      - name: productid
        in: path
        description: Identifier of the Product
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                response:
                  value: |-
                    {
                      "product": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Intel BT200 Motherboard",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.company_controller
    put:
      tags:
      - Company
      summary: Update a Product
      description: Stores a Product
      operationId: update_product
      parameters:
      - name: productid
        in: path
        description: Identifier of the Product
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      requestBody:
        $ref: '#/components/requestBodies/Product'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                response:
                  value: |-
                    {
                      "product": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "Intel BT200 Motherboard",
                      "active": true,
                      "company": "0e8cedd0-ad98-11e6-bf2e-47111ada7c0f"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
    delete:
      tags:
      - Company
      summary: Delete a Product
      description: Deletes a Product
      operationId: delete_product
      parameters:
      - name: productid
        in: path
        description: Identifier of the Product
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
  /company/:
    get:
      tags:
      - Company
      summary: Load the list of Companies
      description: Loads list of Companies
      operationId: search_companies
      parameters:
      - name: size
        in: query
        description: Size of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: integer
        x-example: 10
      - name: page
        in: query
        description: Number of the page to retrieve.
        required: false
        style: form
        explode: true
        schema:
          type: number
        x-example: 1
      - name: sort
        in: query
        description: 'Order in which to retrieve the results. Multiple sort criteria
          can be passed. Example: sort=name ASC,city DESC'
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: name ASC
      - name: name
        in: query
        description: Allows to filter the collections of result by the value of name
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-example: George Street Brewery
      responses:
        "200":
          description: Status 200
          headers:
            X-Page-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Number:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 1
            X-Page-Size:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 25
            X-Total-Count:
              style: simple
              explode: false
              schema:
                type: integer
              x-example: 2
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'
                x-content-type: application/json
              examples:
                response:
                  value: |-
                    [{
                      "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "George Street Brewery",
                      "address":{
                        "street": "2 place de la Defense",
                        "zipcode": "92053",
                        "city": "Paris",
                        "taxid": "29.691.807/0001-87"
                      }
                    }]
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.company_controller
    post:
      tags:
      - Company
      summary: Create a new Company
      description: Adds a Company
      operationId: add_company
      requestBody:
        $ref: '#/components/requestBodies/Company'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              examples:
                response:
                  value: |-
                    {
                      "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "George Street Brewery",
                      "address":{
                        "street": "2 place de la Defense",
                        "zipcode": "92053",
                        "city": "Paris",
                        "taxid": "29.691.807/0001-87"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
  /company/{companyid}:
    get:
      tags:
      - Company
      summary: Load an individual Company
      description: Loads a Company
      operationId: get_company
      parameters:
      - name: companyid
        in: path
        description: Identifier of the Company
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              examples:
                response:
                  value: |-
                    {
                      "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "George Street Brewery",
                      "address":{
                        "street": "2 place de la Defense",
                        "zipcode": "92053",
                        "city": "Paris",
                        "taxid": "29.691.807/0001-87"
                      }
                    }
        "400":
          description: Status 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.company_controller
    put:
      tags:
      - Company
      summary: Update a Company
      description: Stores a Company
      operationId: update_company
      parameters:
      - name: companyid
        in: path
        description: Identifier of the Company
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      requestBody:
        $ref: '#/components/requestBodies/Company'
      responses:
        "200":
          description: Status 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
              examples:
                response:
                  value: |-
                    {
                      "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
                      "name": "George Street Brewery",
                      "address":{
                        "street": "2 place de la Defense",
                        "zipcode": "92053",
                        "city": "Paris",
                        "taxid": "29.691.807/0001-87"
                      }
                    }
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
    delete:
      tags:
      - Company
      summary: Delete a Company
      description: Deletes a Company
      operationId: delete_company
      parameters:
      - name: companyid
        in: path
        description: Identifier of the Company
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      responses:
        "200":
          description: Status 200
      security:
      - HTTP_BASIC: []
      x-openapi-router-controller: swagger_server.controllers.company_controller
components:
  schemas:
    Error:
      required:
      - code
      type: object
      properties:
        code:
          maximum: 599
          minimum: 400
          type: integer
        description:
          type: string
          example: 'Bad query parameter [$size]: Invalid integer value [abc]'
        reasonPhrase:
          type: string
          example: Bad Request
      description: This general error structure is used throughout this API.
      example: |-
        {
          "code": 400,
          "description": "Bad query parameter [$size]: Invalid integer value [abc]",
          "reasonPhrase": "Bad Request"
        }
    Product:
      required:
      - company
      - name
      - product
      type: object
      properties:
        product:
          type: string
          description: Auto-generated primary key field
          example: 0e8ffb10-ad98-11e6-bf2e-47644ada7c0f
        name:
          type: string
          example: Fast Ethernet Adapter 10/100 Mbps
        active:
          type: boolean
          default: true
        company:
          type: string
          description: This property is a reference to a Company
          example: 0e8c9fb0-ad98-11e6-bf2e-47644ada7c0f
      description: A representation of a Product.
      example: |-
        {
          "product": "0e8dd830-ad98-11e6-bf2e-47644ada7c0f",
          "name": "Fast Ethernet Adapter 10/100 Mbps",
          "active": true,
          "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f"
        }
...
    Company:
      required:
      - address
      - company
      - name
      type: object
      properties:
        company:
          type: string
          description: Auto-generated primary key field
        name:
          type: string
        address:
          $ref: '#/components/schemas/Company_address'
      description: A representation of a Company.
      example: |-
        {
          "company": "0e8cedd0-ad98-11e6-bf2e-47644ada7c0f",
          "name": "George Street Brewery",
          "address":{
            "street": "2 place de la Defense",
            "zipcode": "92053",
            "city": "Paris",
            "taxid": "29.691.807/0001-87"
          }
        }
    Company_address:
      required:
      - city
      - street
      - taxId
      - zipcode
      type: object
      properties:
        street:
          type: string
        street2:
          type: string
        city:
          type: string
        zipcode:
          pattern: '[0-9]*'
          type: string
        taxId:
          type: string
          example: 29.691.803/0001-80
  requestBodies:
...
    Product:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Product'
      required: true
...
    Company:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Company'
      required: true
  securitySchemes:
    HTTP_BASIC:
      type: http
      description: All GET methods are public, meaning that *you can read all the
        data*. Write operations require authentication and therefore are forbidden
        to the general public.
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_HTTP_BASIC